[model] use hibernate and refine type and wording
