+++ src/main/java/app/controller/BookController.java
@@ -7,19 +7,19 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import app.ServicePath;
-import app.repostory.Author;
-import app.repostory.Book;
-import app.repostory.Publisher;
-import app.repostory.Translator;
+import app.model.Author;
+import app.model.Book;
+import app.model.Publisher;
+import app.model.Translator;
 
 @RestController
 public class BookController {
 
 	@GetMapping(value = ServicePath.BOOK_COLLECTION, produces = MediaType.APPLICATION_JSON_VALUE)
 	public Book getBook() {
-		return new Book(1, "Book Name", new Author(1, "Author. Name.", null),
-				new Translator(1, "Translator. Name.", null), "ISBN-CODE-123-456-789",
-				new Publisher(1, "Publisher. Name.", null), new Date(), 100.0);
+		return new Book("Book Name", new Author("Author. Name."),
+				new Translator("Translator. Name."), "ISBN-CODE-123-456-789",
+				new Publisher("Publisher. Name."), new Date(), 100.0);
 	}
 
 }
